src/main/java/com/gabriel/servicos/LocacaoService.java
package com.gabriel.servicos;

import static com.gabriel.utils.DataUtils.adicionarDias;

import java.util.Date;

import com.gabriel.entidades.Filme;
import com.gabriel.entidades.Locacao;
import com.gabriel.entidades.Usuario;

public class LocacaoService {
	
	public Locacao alugarFilme(Usuario usuario, Filme filme) {
		Locacao locacao = new Locacao();
		locacao.setFilme(filme);
		locacao.setUsuario(usuario);
		locacao.setDataLocacao(new Date());
		locacao.setValor(filme.getPrecoLocacao());
		
		//Entrega no dia seguinte
		Date dataEntrega = new Date();
		dataEntrega = adicionarDias(dataEntrega, 1);
		locacao.setDataRetorno(dataEntrega);
		
		//Salvando a locacao...	
		//TODO adicionar método para salvar
		
		return locacao;
	}
}